+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,26 @@
 # alpha-test-app
 
+## 0.0.1-alpha.32
+
+### Patch Changes
+
+- Updated dependencies [3198059]
+  - sit-onyx@1.0.0-alpha.35
+
+## 0.0.1-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [a8ad4ff]
+  - sit-onyx@1.0.0-alpha.34
+
+## 0.0.1-alpha.30
+
+### Patch Changes
+
+- Updated dependencies [ab2fbb8]
+  - sit-onyx@1.0.0-alpha.33
+
 ## 0.0.1-alpha.29
 
 ### Patch Changes