+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,23 @@
 # sit-onyx
 
+## 1.0.0-alpha.35
+
+### Minor Changes
+
+- 3198059: feat: add property `skeleton` to `OnyxCheckboxGroup` and `OnyxRadioButtonGroup`
+
+## 1.0.0-alpha.34
+
+### Patch Changes
+
+- a8ad4ff: fix(OnyxButton): add active state
+
+## 1.0.0-alpha.33
+
+### Minor Changes
+
+- ab2fbb8: feat: add basic OnyxInput component
+
 ## 1.0.0-alpha.32
 
 ### Minor Changes