+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-alpha.37
+
+### Patch Changes
+
+- Updated dependencies [3198059]
+  - sit-onyx@1.0.0-alpha.35
+
+## 1.0.0-alpha.36
+
+### Patch Changes
+
+- Updated dependencies [a8ad4ff]
+  - sit-onyx@1.0.0-alpha.34
+
+## 1.0.0-alpha.35
+
+### Patch Changes
+
+- Updated dependencies [ab2fbb8]
+  - sit-onyx@1.0.0-alpha.33
+
 ## 1.0.0-alpha.34
 
 ### Patch Changes